style: 피드백 반영하여 클래스 구조 수정
